move methods to resolve type issue
