+++ qiskit_nature/problems/second_quantization/lattice/models/ising_model.py
@@ -18,6 +18,7 @@ from typing import Optional
 import numpy as np
 
 from qiskit_nature.operators.second_quantization import SpinOp
+from qiskit_nature.problems.second_quantization.lattice.lattices import Lattice
 
 from .lattice_model import LatticeModel
 
@@ -30,6 +31,49 @@ class IsingModel(LatticeModel):
     def coupling_matrix(self) -> np.ndarray:
         """Return the coupling matrix."""
         return self.interaction_matrix()
+
+    @classmethod
+    def uniform_parameters(
+        cls,
+        lattice: Lattice,
+        uniform_interaction: complex,
+        uniform_onsite_potential: complex,
+    ) -> "IsingModel":
+        """Set a uniform interaction parameter and on-site potential over the input lattice.
+
+        Args:
+            lattice: Lattice on which the model is defined.
+            uniform_interaction: The interaction parameter.
+            uniform_onsite_potential: The on-site potential.
+
+        Returns:
+            The Lattice model with uniform parameters.
+        """
+        return cls(
+            cls._generate_lattice_from_uniform_parameters(
+                lattice, uniform_interaction, uniform_onsite_potential
+            )
+        )
+
+    @classmethod
+    def from_parameters(
+        cls,
+        interaction_matrix: np.ndarray,
+    ) -> "IsingModel":
+        """Return the Hamiltonian of the Lattice model
+        from the given interaction matrix and on-site interaction.
+
+        Args:
+            interaction_matrix: A real or complex valued square matrix.
+
+        Returns:
+            LatticeModel: The Lattice model generated from the given interaction
+                matrix and on-site interaction.
+
+        Raises:
+            ValueError: If the interaction matrix is not square matrix, it is invalid.
+        """
+        return cls(cls._generate_lattice_from_parameters(interaction_matrix))
 
     def second_q_ops(self, display_format: Optional[str] = None) -> SpinOp:
         """Return the Hamiltonian of the Ising model in terms of `SpinOp`.
